Refactoring for SortValidator
- SortValidator now can be used for all entities
