+++ src/main/java/com/rj/ecommerce_backend/domain/sortingfiltering/CategorySortField.java
@@ -1,0 +1,18 @@
+package com.rj.ecommerce_backend.domain.sortingfiltering;
+
+public enum CategorySortField implements SortableField {
+    ID("id"),
+    NAME("name");
+
+    private final String fieldName;
+
+    CategorySortField(String fieldName) {
+        this.fieldName = fieldName;
+    }
+
+    @Override
+    public String getFieldName() {
+        return fieldName;
+    }
+
+}
